refactored common Kafka2 and Kafka3 code
