--- pinot-plugins/pinot-stream-ingestion/pinot-kafka-2.0/src/main/java/org/apache/pinot/plugin/stream/kafka20/KafkaConsumerPartitionLag.java
@@ -1,41 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.plugin.stream.kafka20;
-
-import org.apache.pinot.spi.stream.PartitionLagState;
-
-
-public class KafkaConsumerPartitionLag extends PartitionLagState {
-  private final String _recordsLag;
-  private final String _availabilityLagMs;
-
-  public KafkaConsumerPartitionLag(String recordsLag, String availabilityLagMs) {
-    _recordsLag = recordsLag;
-    _availabilityLagMs = availabilityLagMs;
-  }
-
-  public String getRecordsLag() {
-    return _recordsLag;
-  }
-
-  @Override
-  public String getAvailabilityLagMs() {
-    return _availabilityLagMs;
-  }
-}
