+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-3.0/src/main/java/org/apache/pinot/plugin/stream/kafka30/KafkaConsumerFactory.java
@@ -18,6 +18,7 @@
  */
 package org.apache.pinot.plugin.stream.kafka30;
 
+import org.apache.pinot.plugin.stream.kafka.KafkaConfigBackwardCompatibleUtils;
 import org.apache.pinot.spi.stream.PartitionGroupConsumer;
 import org.apache.pinot.spi.stream.PartitionGroupConsumptionStatus;
 import org.apache.pinot.spi.stream.StreamConfig;
