+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-base/src/main/java/org/apache/pinot/plugin/stream/kafka/KafkaConsumerPartitionLag.java
@@ -16,22 +16,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.plugin.stream.kafka30;
+package org.apache.pinot.plugin.stream.kafka;
 
-import java.util.Map;
-import javax.annotation.Nullable;
-import org.apache.pinot.spi.data.readers.GenericRow;
-import org.apache.pinot.spi.stream.StreamMessageMetadata;
+import org.apache.pinot.spi.stream.PartitionLagState;
 
-// TODO: Make it a util class
-public class KafkaStreamMessageMetadata extends StreamMessageMetadata {
-  public static final String METADATA_OFFSET_KEY = "offset";
-  public static final String RECORD_TIMESTAMP_KEY = "recordTimestamp";
-  public static final String METADATA_PARTITION_KEY = "partition";
 
-  @Deprecated
-  public KafkaStreamMessageMetadata(long recordIngestionTimeMs, @Nullable GenericRow headers,
-      Map<String, String> metadata) {
-    super(recordIngestionTimeMs, headers, metadata);
+public class KafkaConsumerPartitionLag extends PartitionLagState {
+  private final String _recordsLag;
+  private final String _availabilityLagMs;
+
+  public KafkaConsumerPartitionLag(String recordsLag, String availabilityLagMs) {
+    _recordsLag = recordsLag;
+    _availabilityLagMs = availabilityLagMs;
+  }
+
+  public String getRecordsLag() {
+    return _recordsLag;
+  }
+
+  @Override
+  public String getAvailabilityLagMs() {
+    return _availabilityLagMs;
   }
 }
